Move tests for immutable arbor into its own file
